C2.6.2. Refactoring - Template Decorator for validator. Unstable
