+++ chapter_002/src/test/java/tracker/model/ItemTest.java
@@ -1,4 +1,4 @@
-package tracker;
+package tracker.model;
 
 import org.junit.Test;
 import tracker.model.Item;
